Guard against null to dirname
Since Coffeescript traces are useless it's impossible to know which of
the dirname calls was triggering the issue, so guard against it for all
of them. If the given path was null in these cases we are probably out
of luck anyway, but in case something can be found set it to __dirname.